Extract table names and data for composite IODs
Parse through the DICOM standard and extract all tables and their
corresponding table names in chapter A. The data from each table
is made available as a list. This list is not yet formatted, and each
table may have a different number of columns. The parser does not yet
separate out different types of tables in chapter A, so no structure is
guaranteed for the table data.

The parser.py script is a first draft, so it makes calls to subprocess
that will later be moved to a Makefile.

Finally, include the HTML file for the DICOM standard for testing
purposes.