Add recursive macro searching for modules
Find and follow all include links to other macro tables and insert the
attributes into the given module. The program looks for the "include"
keyword in the first column of every row in a table and follows the
link, if present.

Also adds a safeguard against infinitely recursive includes, which
occur in the SR Document Content module.