Add rowspan expansion and switch to lists
Expanding rowspan attributes in a list-of-list is now implemented, and
the table data is now being stored in a list of lists instead of a
panda dataframe. This decision was made because the panda dataframe had
a large amount of syntactical overhead and did not allow for many
common row operations.

Also, unit tests for all of these functions have been added. Next commit
will be using these components to reconstruct the rough draft parser's
functionality.