Move Makefile and organize, improve unit tests
Moved Makefile to top directory and added easier keywords to build
portions of the project. Also added a "test" rule to run all of the
pytest functions while avoiding the ones in the virtual environment
folder.

Also, made a more comprehensive unit test for the input file that
checks all DICOM tags in the file and compares them to the JSON
representation.