Improve handling of DICOM sequences in JSON
Now, DICOM sequences are considered to be arrays of dicts (same
behavior as pydicom). The string representation of the value is what is
stored in the JSON file.

Also, refactored some of functions in the input file parsing to make
the code more readable and logical, and added more unit tests checking
the behavior of JSON conversion of a DICOM sequence.