Add VR, VM, keyword to attrs, JSON keys lowercase
Added new script and Makefile rules to add the VR, VM, and keyword
fields to the attribute JSON file. Also, modified JSON keys to all be
lowercase and replaced all spaces with underscores.

This commit addresses gh-9. Still to be done is adding the composite
IOD description, but all other fields are present.