Rename functions and add tests for parent ids
Renamed functions in parse_lib.py as well as added unit tests covering
the finding and addition of parent ids to attributes and the
normalization process.

Finally, also moved the top level function in many of the scripts into
the __main__ block in order to reduce the number of small functions in
these files.