Modify schema to module <-> (sequence, attribute)
Keep track of the sequence that an attribute came from when identifying
its relationship to the module-- otherwise, a lot of macros get called
multiple times in different sequences and they override each other in
the table.

Also, adding some end-to-end tests. Still a work in progress, but
the tests confirm that all CIODs and attributes are accounted for.