Fix retired behavior and type normalization
The retired field for an attribute is now set by a regex match to the
term "RET" instead of just the presence of a string in the final
column. This was done because the last column in PS3.6, although
labeled as "retired", actually holds multiple values.

Also, fixed an issue where attribute normalization would assign the
first type that it found in the module-attribute relationship table.
This has been remedied by continuing to search until a non-None type
has been found, and assigning None if there is no matching type.

Finally, added more end-to-end tests that check the behavior of these
tweaks on the output by testing specific attributes from the standard.