Move makefile to parse, break up tests
Changing the organization scheme of the repository to enforce
separation between the different stages of the project. The parsing
portion of the project now has its own makefile, directory of
applicable tests, and Dockerfile. A similar process has been done for
the api directory, and the two are linked by a top level Jenkinsfile
which specifies the CI/CD approach for the project.