Remoeve pandas as a dependency
It really was taking a long time to build, and after some reconsideration,
I think it isn't worth it to use it.  After all, at the moment we are
using it for only a small amount of functionality, and it isn't clear
how we will use it in the immediate future.