Add ugly hook to skip rows if necessary
The need for doing this, indicates that the current structure is too
tightly tied to the specific tables being parsed.  Really, a lot of the
functions that are 2 or 3 levels deep should get pulled up a couple
levels and out of the general `parse_lib`, however there is not time for
this now.

Whenever we need to come back and add more parsing functionality is when
we should do this.

NOTE: The invocation of this prob spilled into the previous commit