Repair test failures, remove useless steps in make
Repair some of the end-to-end test failures that happened during the
refactor.

Also, a big realization was that the module_attribute relationship
table only depends on the information in PS3.3, and the attribute table
depends only on the information in PS3.6. Therefore, there's actually
no need to merge the tables together at all. This is a bit of a pain to
figure out this late in the game...

There is still one test that is not passing, and it unearthed a
frustrating bug that still needs to be tracked down. For some reason,
not all attributes are being picked up in a module (see
end_to_end_test.py, line 167 for an example). This is occurring at the
very first parsing stage, so something that happened during the
refactor at the first parse level has changed behavior slightly. I've
explored this bug some, but it needs more investigation before merging
with master.