Fix issues in dicomTableTree, change JSON output
Resolved issues in dicomTableTree that were showing up because of
inconsistent path delimeters (sometimes "/" and sometimes ":", resolved
by sticking with ":").

Also, modified the attributes.json file to use the ID version of the
tag (i.e. "00010001") as the unique key instead of the formatted tag
(i.e. "(0001,0001)"). This made it easier to adjust some of the
filtering steps in dicomTableTree.

As a performance note, it seems that loading the relevant attributes
onClick is the speed bottleneck. Future work should include finding
ways to optimize these lookups.