Update unit tests, add attribute cleaning fcn
Updated unit tests in the parser after changes made to extract raw HTML
for the description fields. More work should be done to clarify some of
these unit tests-- while the tests are all checking defined and expected
behavior, there is not much explanation as to why the sample data sets
have some elements that are raw HTML and others that are parsed text.

Also, added an attribute cleaning function to remove attributes from
HTML nodes (with the exception of `href` tags) and added unit tests to
check the functionality.