Add TODO for uncovered bug
Accidentally uncovered a bug in the href link generation for references
to sections other than the current section (i.e. `part03.html`). As
described in the comment:

> hrefs to other sections get incorrectly assembled
> i.e. If we are in part03.html:
> http://dicom.nema.org/medical/dicom/current/output/html/part03.htmlpart06.html#PS3.6

Will resolve this bug before merging with master, but I want to keep
moving forward with the goal of this branch before I get sidetracked bug
squashing.