Begin building skeleton of extra parsing func
This commit is still a work in progress.

A rough (almost functional) outline of the extra parsing step is
implemented and added to the Makefile workflow. TODOs have been added
where appropriate.

Note: the `get_section_html` function in `parse_extra_sections.py` has a
visually very unappealing one-liner that is strongly coupled to the
structure of the HTML. All HTML sections comply with this pattern, but
I placed it in a function since it syntactically looks very bizarre.