Change JSON structure, parse later in build
Changed the JSON structure of the extra parsed sections to be key-value
pairs, where the key is the anchor link text and the value is the HTML
string of the referenced section.

Also, moved the parsing to after normalization so that the data is the
most refined. It also just makes the parsing simpler.