Fix "http://" error in sitemaps, file server
Apparently, not only was Google not seeing our sitemaps as actual URLs
because of the missing "http://" prefix, but it also couldn't get the
files to begin with because expressjs was redirecting to a 404 page!
Added a handle to our express server to manage sitemap requests. Also
modified the sitemap generation code to add the "http://" prefix.