Reduce sitemap generation time from ~ 10min -> 1s
The sitemap generation was the slowest step in our build process.

The process was slow because of two nested for loops that iterated
through the entire ciod-to-module and module-to-attribute dicts many
more times than was necessary.

On my computer this step was taking on the order of 20 minutes, and it
nome takes less of a second.

The key to the speed improvement was to re-organize the data into a tree
in memory, and to then write out its contents.

This commit also:

- Adds a couple unit tests
- Fixes the URLs in the sitemap index (they did not include "http://")
- Splits the code up into more logical pieces

Still need to:

- Add a check to complain if a sitemap is too larger
- Validate the output in Google

Manual verification indicated that the files were in good shape.