Add module descriptions from DICOM standard
Parsed the header information above each module table and converted it
into the module description. This works pretty darn well, but a small
hiccup is that most descriptions begin with a link to the table that
describes the module in the DICOM standard. We'll probably want to
deactivate this link and most likely change the text, because otherwise
it could be pretty confusing for users to see the description reference
a table that doesn't exist.