Finish the extract_modules_with_attributes stage
There is a known parsing error due to an inconsistency in the DICOM
website's table spanning conventions. However, this error will not be an
issue since it is occurring in a row that will later be expanded into a
macro; therefore, it's not worth it to develop a one-off error
detection. This only happens on one particular table of the standard, so
it's a very specialized case.