WIP: Begin preprocess stage, modify macro JSON
The JSON structure for the `extract_macros.py` stage was modified so
that each macro was accessible by its HTML key (i.e. the key used in the
URL fragment). This made finding macros an O(1) process instead of O(n),
which will be very useful considering the number of times we load a
macro.

Before moving to the `expand_heirarchy` substage, heirarchy markers (the
">" characters before an attribute name) need to be added to all the
newly added macro attributes.