WIP: begin adding hierarchy markers
Due to the way text replacement works in BeautifulSoup, we can't simply
replace the `string` attribute of a given parsed tag-- it will actually
overwrite all child tags of the element you are editing, not just the
text.

Next steps: figure out if we even need to keep the HTML intact or if we
can go straight to a string representation.