Delete old stages and clean parse_lib
Deleted the old parsing stages that have been redesigned and replaced.
In the process, a lot of old functions from `parse_lib.py` were able to
be deleted. There is still more deleting to be had after the refactor is
complete, but this is to clear space so all the files don't become
confusing.