Update proposal with change in sections JSON
In order to support referenced material from other sections, an extra
level of information was added where each section of the dicom standard
acted as a key to a dictionary containing all of its subsections. This
worked out well since it also mimicks the structure of the URL fragment.

    <BASE_DICOM_URL>partXX.html#my_section_id

maps to

    sections['partXX.html']['my_section_id']