Remove unnecessary reference util module
The references_utils module was only used by postprocess_add_references,
hence it seemed like it was not necessary to split it out into a
separate module.  Placing the code in a separate module just made the
code more difficult to read, because you have to move over to the other
file to look at the code.  Also, it is misleading, because the fact that
it is in a separate file implies it is used elsewhere.