Add title field to references, fix html clean bug
A title field was necessary for each reference so that the dicom browser
application had a header to load for every reference even before the
HTML content was received.

Also, the conversion of the reference anchor tag to a `span` was not
occurring because the modified parse object was not being saved back to
the description field. This was fixed and a few functions were clarified
in the process.

Finally, this commit also removes an inefficiency in the
`postprocess_save_references.py` stage where each pair was being parsed
again, even though all the reference information had already been
extracted in the previous stage.