Fix macro nesting error when counting markers
Some attribute name cells have newlines for formatting. These newlines
break apart the different hierarchy markers (">") in the text
representation of the HTML. In the old implementation of the hierarchy
level counter, the regex '^(>+)' was used to count the number of
markers; however, this match ends at the first newline or break in
between markers.

This was fixed by removing newlines and spaces before counting the
number of markers in a name.

In addition to this, a unit test was added to explicitly check macro
expansion on a mock table.