Support short URLs of non-section references
This required a change in how references were parsed, since many
non-section references (i.e. images) did not retain their enclosing
section information after parsing. However, the enclosing section
information is required in order to properly reconstruct the small DICOM
URL of the reference.

A final postprocessing stage had to be added for
`dist/module_to_attributes.json` in order for the sourceUrl attribute to
be properly updated. This is because the
`postprocess_mark_references.py` stage only knows a reference by its
fragment that is present in the description of a module or attribute;
these fragments do not contain any indication as to the enclosing
section ID. Therefore, the final URL cannot be constructed without the
complete reference information being passed in at a later stage.