Revise map usage, naming, and SVG parsing
Removed any calls to `map` that generated side effects and replaced with
explicit loops in order to make the side effects more clear. Also
removed any unneccesary `list` calls on iterables.

Also updated SVG parsing so that it remains at the abstraction layer of
an SVG and not the underlying HTML tag. Part of this required more
intelligent parsing of object tags to look only for those typed with
`svg`; after this, functions were renamed to reflect this change.

Also, some small renaming changes to enhance readability.